+++ hack/testdata/levee/levee-config.yaml
@@ -92,9 +92,6 @@ Sources:
 - PackageRE: "k8s.io/kubernetes/pkg/kubelet/client"
   TypeRE: "KubeletClientConfig"
   FieldRE: "BearerToken"
-- PackageRE: "k8s.io/kubernetes/pkg/kubelet/cri/streaming"
-  TypeRE: "cacheEntry"
-  FieldRE: "token"
 
 # The following fields are tagged in #96005
 - PackageRE: "k8s.io/api/authentication/v1"
