+++ pkg/client/tests/portfoward_test.go
@@ -35,7 +35,7 @@ import (
 	restclient "k8s.io/client-go/rest"
 	. "k8s.io/client-go/tools/portforward"
 	"k8s.io/client-go/transport/spdy"
-	"k8s.io/kubernetes/pkg/kubelet/cri/streaming/portforward"
+	"k8s.io/kubelet/pkg/cri/streaming/portforward"
 )
 
 // fakePortForwarder simulates port forwarding for testing. It implements
